+++ packages/ember-rdfa-editor/CHANGELOG.md
@@ -1,5 +1,93 @@
 # @lblod/ember-rdfa-editor
 
+## 13.1.1
+
+### Patch Changes
+
+- [#1352](https://github.com/lblod/ember-rdfa-editor/pull/1352) [`02b7d3a`](https://github.com/lblod/ember-rdfa-editor/commit/02b7d3ad477003c78388ba20526e4d05518ab4a1) Thanks [@piemonkey](https://github.com/piemonkey)! - Upgrade some minor or patch versions of dependencies
+
+- [#1349](https://github.com/lblod/ember-rdfa-editor/pull/1349) [`d68bae9`](https://github.com/lblod/ember-rdfa-editor/commit/d68bae949baa71aa5c8b919b86c02a09290e1df5) Thanks [@piemonkey](https://github.com/piemonkey)! - Remove deprecated internal use of AuContent within AuCard components
+
+## 13.1.0
+
+### Minor Changes
+
+- [#1348](https://github.com/lblod/ember-rdfa-editor/pull/1348) [`a09a6f4`](https://github.com/lblod/ember-rdfa-editor/commit/a09a6f41fd4dbce52f09878094423f1781efba11) Thanks [@piemonkey](https://github.com/piemonkey)! - Only apply RDFa block and inline styling when within the editor, to avoid extra borders in previewed content
+
+## 13.0.0
+
+### Major Changes
+
+- [#1346](https://github.com/lblod/ember-rdfa-editor/pull/1346) [`51a2729`](https://github.com/lblod/ember-rdfa-editor/commit/51a27291cb902bbe8597d0b2d7d020672eaa9694) Thanks [@lagartoverde](https://github.com/lagartoverde)! - Remove list tracking plugin and add a onChange property on list nodes
+
+## 12.15.1
+
+### Patch Changes
+
+- [#1345](https://github.com/lblod/ember-rdfa-editor/pull/1345) [`c051aad`](https://github.com/lblod/ember-rdfa-editor/commit/c051aad98a86fa12c495a4ed8b5ec8cf9420f326) Thanks [@lagartoverde](https://github.com/lagartoverde)! - Fix ListTracking and onChanged plugins were looping
+
+## 12.15.0
+
+### Minor Changes
+
+- [#1340](https://github.com/lblod/ember-rdfa-editor/pull/1340) [`7e6de75`](https://github.com/lblod/ember-rdfa-editor/commit/7e6de75995d22a8ba62472de0fd9f9b3fdf617c1) Thanks [@abeforgit](https://github.com/abeforgit)! - Add new semantic html string comparison assertion
+
+- [#1343](https://github.com/lblod/ember-rdfa-editor/pull/1343) [`10127b5`](https://github.com/lblod/ember-rdfa-editor/commit/10127b51e87b8dfa54d28a80b77140906bef19ad) Thanks [@elpoelma](https://github.com/elpoelma)! - Add support for model migrations to `block_rdfa` node
+
+- [#1340](https://github.com/lblod/ember-rdfa-editor/pull/1340) [`7e6de75`](https://github.com/lblod/ember-rdfa-editor/commit/7e6de75995d22a8ba62472de0fd9f9b3fdf617c1) Thanks [@abeforgit](https://github.com/abeforgit)! - Move all dependencies to pnpm catalogs
+
+### Patch Changes
+
+- [#1340](https://github.com/lblod/ember-rdfa-editor/pull/1340) [`7e6de75`](https://github.com/lblod/ember-rdfa-editor/commit/7e6de75995d22a8ba62472de0fd9f9b3fdf617c1) Thanks [@abeforgit](https://github.com/abeforgit)! - bump dependencies
+
+## 12.14.0
+
+### Minor Changes
+
+- [#1341](https://github.com/lblod/ember-rdfa-editor/pull/1341) [`85f3652`](https://github.com/lblod/ember-rdfa-editor/commit/85f3652dbcce2f4861f0c2384466503bc965c899) Thanks [@elpoelma](https://github.com/elpoelma)! - `transactionCombinator`: apply manual selection changes performed in the transaction-monads
+
+- [#1341](https://github.com/lblod/ember-rdfa-editor/pull/1341) [`85f3652`](https://github.com/lblod/ember-rdfa-editor/commit/85f3652dbcce2f4861f0c2384466503bc965c899) Thanks [@elpoelma](https://github.com/elpoelma)! - `transactionCombinator`: apply all steps of the resulting transactions coming from `State.apply`
+
+## 12.13.0
+
+### Minor Changes
+
+- [#1338](https://github.com/lblod/ember-rdfa-editor/pull/1338) [`18d97bc`](https://github.com/lblod/ember-rdfa-editor/commit/18d97bcea68a87782ff5457c7f7396ab77973430) Thanks [@piemonkey](https://github.com/piemonkey)! - Add preliminary support for data model migrations when parsing documents
+
+## 12.12.1
+
+### Patch Changes
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`a5bb5eb`](https://github.com/lblod/ember-rdfa-editor/commit/a5bb5eb4321482292aed4365fec57053c1da7f4e) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-inputrules` to version 1.5.0
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`eb306d8`](https://github.com/lblod/ember-rdfa-editor/commit/eb306d83913640dfa04e595ef4d0faa2802bdb21) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-schema-list` to version 1.5.1
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`5992dc4`](https://github.com/lblod/ember-rdfa-editor/commit/5992dc4bc2fcb281ecc9060d219ba3a2e02940c6) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-commands` to version 1.7.1
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`8dd20bf`](https://github.com/lblod/ember-rdfa-editor/commit/8dd20bfd7915f5fcc18b01eb80bffe12d1619f3c) Thanks [@elpoelma](https://github.com/elpoelma)! - Deprecate `toggleMarkAddFirst` command, use the the `toggleMark` command with the `removeWhenPresent` option instead.
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`8f5c620`](https://github.com/lblod/ember-rdfa-editor/commit/8f5c6206ce7394ef8522b1ca511e24f2a0cbc204) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-transform` dependency to version 1.10.4
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`d59e43b`](https://github.com/lblod/ember-rdfa-editor/commit/d59e43b6f30040340bbc41cd4b8e0af5beed70b3) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-view` dependency to version 1.40.1
+
+- [#1336](https://github.com/lblod/ember-rdfa-editor/pull/1336) [`61ca675`](https://github.com/lblod/ember-rdfa-editor/commit/61ca675af52b1c0a6c4caa76dda460b1b30014df) Thanks [@elpoelma](https://github.com/elpoelma)! - Update `prosemirror-model` to version 1.25.3
+
+- [#1337](https://github.com/lblod/ember-rdfa-editor/pull/1337) [`f7533b0`](https://github.com/lblod/ember-rdfa-editor/commit/f7533b0555925143815be5134cf5a069d9fac202) Thanks [@abeforgit](https://github.com/abeforgit)! - fix parsing of resourcenodes with contentLiterals
+
+## 12.12.0
+
+### Minor Changes
+
+- [#1331](https://github.com/lblod/ember-rdfa-editor/pull/1331) [`f68bff4`](https://github.com/lblod/ember-rdfa-editor/commit/f68bff4d9b031261fcf9316cb875b405e47baeee) Thanks [@elpoelma](https://github.com/elpoelma)! - Add explicit support for ember-source version 6.x
+
+- [#1334](https://github.com/lblod/ember-rdfa-editor/pull/1334) [`0945144`](https://github.com/lblod/ember-rdfa-editor/commit/09451449a9bcde555ea2c331c5672f2e7c552948) Thanks [@piemonkey](https://github.com/piemonkey)! - Add context to RDFa visualiser components that says if a resource is at the top level so the display can be changed accordingly
+
+### Patch Changes
+
+- [#1333](https://github.com/lblod/ember-rdfa-editor/pull/1333) [`ba2ebd0`](https://github.com/lblod/ember-rdfa-editor/commit/ba2ebd045613efb37db16122aa74a49e0d4859ef) Thanks [@elpoelma](https://github.com/elpoelma)! - rdfa-visualizer: fully hide resource node when it is marked as `hidden: true` in the visualizer configuration.
+
+- [#1332](https://github.com/lblod/ember-rdfa-editor/pull/1332) [`f73d5f1`](https://github.com/lblod/ember-rdfa-editor/commit/f73d5f144b9516e1c069eb4c02e559abff5ebaa5) Thanks [@elpoelma](https://github.com/elpoelma)! - rdfa-visualizer: correctly render property objects using `ConfigurableRdfaDisplay` instead of `ResourceInfo`
+
 ## 12.11.0
 
 ### Minor Changes
@@ -53,7 +141,6 @@
 - [#1317](https://github.com/lblod/ember-rdfa-editor/pull/1317) [`b24057e`](https://github.com/lblod/ember-rdfa-editor/commit/b24057e7d3d08afa73a060ed2c8c52e2c44c47b2) Thanks [@elpoelma](https://github.com/elpoelma)! - relationship-editor: add utility `combineConfigs` function
 
 - [#1310](https://github.com/lblod/ember-rdfa-editor/pull/1310) [`7886d5d`](https://github.com/lblod/ember-rdfa-editor/commit/7886d5d3c35e7c65fadc8c958c69244739ac804b) Thanks [@elpoelma](https://github.com/elpoelma)! - Updates to the `OptionGenerator` interfaces
-
   - The `subjectOptionGenerator`, `predicateOptionGenerator` and `objectOptionGenerator` have been collected into a single `optionGeneratorConfig` argument. This makes it easier to provide sample/default configs as a single variable/export.
   - `PredicateOption`: if the `direction` is 'property', an optional `allowFreeTextTarget` may be passed. This allows end-users to type in a free-text literal as object when that specific predicate option is selected.
   - the `PredicateOptionGenerator` now also optionally receives a `Direction` ('property'/'backlink') as filter input.
@@ -66,7 +153,6 @@
 - [#1321](https://github.com/lblod/ember-rdfa-editor/pull/1321) [`f588378`](https://github.com/lblod/ember-rdfa-editor/commit/f5883786a460ae0f1ac95cfd5f6b3d956afb2cba) Thanks [@piemonkey](https://github.com/piemonkey)! - Fix bug with drag-and-drop for ember-nodes such as snippets
 
 - [#1310](https://github.com/lblod/ember-rdfa-editor/pull/1310) [`7886d5d`](https://github.com/lblod/ember-rdfa-editor/commit/7886d5d3c35e7c65fadc8c958c69244739ac804b) Thanks [@elpoelma](https://github.com/elpoelma)! - Seperate rdfa-tools into seperate components:
-
   - `attribute-editor`
   - `debug-info`
   - `doc-imported-resource-editor`
@@ -86,7 +172,6 @@
 ### Minor Changes
 
 - [#1320](https://github.com/lblod/ember-rdfa-editor/pull/1320) [`6b74bea`](https://github.com/lblod/ember-rdfa-editor/commit/6b74beafb484244bde27f17a2c45ac3eb036fcef) Thanks [@elpoelma](https://github.com/elpoelma)! - Introduction of two new ember-node options:
-
   - `domClassNames`: classnames to apply to the nodeview which is created
   - `contentDomClassNames`: classnames to apply to the content node within the nodeview
 
@@ -253,19 +338,16 @@
 ### Minor Changes
 
 - [#1256](https://github.com/lblod/ember-rdfa-editor/pull/1256) [`e627cfa`](https://github.com/lblod/ember-rdfa-editor/commit/e627cfadd10ad5b50d2d1363f31288f60921d38d) Thanks [@elpoelma](https://github.com/elpoelma)! - `collapsible` component: add option to manually control toggled state through arguments:
-
   - `expanded` (optional, default: `false`)
   - `onToggle` (optional)
 
   The `expandedInitally` argument has been marked as deprecated, and should be replaced by the `expanded` argument
 
 - [#1256](https://github.com/lblod/ember-rdfa-editor/pull/1256) [`8fabad2`](https://github.com/lblod/ember-rdfa-editor/commit/8fabad240e7c4a4895f1b1a00141ebe2ff479f5a) Thanks [@elpoelma](https://github.com/elpoelma)! - `AttributeEditor` component, add arguments to control `expanded` state:
-
   - `onToggle` (optional)
   - `expanded` (optional, default: `true`)
 
 - [#1256](https://github.com/lblod/ember-rdfa-editor/pull/1256) [`5de517b`](https://github.com/lblod/ember-rdfa-editor/commit/5de517b6255dc9bae7499e649944feb25b29b609) Thanks [@elpoelma](https://github.com/elpoelma)! - `DebugInfo` component, add arguments to control `expanded` state:
-
   - `onToggle` (optional)
   - `expanded` (optional, default: `true`)
 
@@ -300,21 +382,18 @@
 - [#1241](https://github.com/lblod/ember-rdfa-editor/pull/1241) [`1517357`](https://github.com/lblod/ember-rdfa-editor/commit/151735798a51452e2318d2393a1385b87f354eda) Thanks [@elpoelma](https://github.com/elpoelma)! - **Adjustments to the `@editorOptions` argument of the `editor-container` component**
 
   The following (legacy) properties are no longer supported:
-
   - `@editorOptions.showRdfaHover`
   - `@editorOptions.showRdfa`
   - `@editorOptions.showRdfaHighlight`
   - `@editorOptions.editRdfa`
 
   The following `@editorOptions` properties have been added:
-
   - `@editorOptions.showSidebarLeft` (default: `true`)
   - `@editorOptions.showSidebarRight` (default: `true`)
 
   These options replace the `@hideSidebar` argument, which is no longer supported.
 
 - [#1241](https://github.com/lblod/ember-rdfa-editor/pull/1241) [`1517357`](https://github.com/lblod/ember-rdfa-editor/commit/151735798a51452e2318d2393a1385b87f354eda) Thanks [@elpoelma](https://github.com/elpoelma)! - **Removal of the following SASS stylesheets**
-
   - `_c-annotation.scss`
   - `_c-annotation-content-en.scss`
   - `_c-annotation-content-nl.scss`
@@ -451,12 +530,10 @@
 ### Minor Changes
 
 - [#1224](https://github.com/lblod/ember-rdfa-editor/pull/1224) [`e37b476`](https://github.com/lblod/ember-rdfa-editor/commit/e37b476a8683f8bd071760e907e7813dfbaad107) Thanks [@elpoelma](https://github.com/elpoelma)! - Simply gap-cursor styling:
-
   - Add small margins
   - Remove unnecessary `position: absolute` and `:after` css rules
 
 - [#1224](https://github.com/lblod/ember-rdfa-editor/pull/1224) [`2cd7ca5`](https://github.com/lblod/ember-rdfa-editor/commit/2cd7ca54ba0fd89662cc416272170328694fca2a) Thanks [@elpoelma](https://github.com/elpoelma)! - Rework `block_rdfa` UX/UI:
-
   - Introduction of `BlockRDFaView` nodeview
   - Addition of a descriptive `label` component to the `block_rdfa` view.
 
@@ -613,7 +690,6 @@
 ### Minor Changes
 
 - [#1195](https://github.com/lblod/ember-rdfa-editor/pull/1195) [`1b7a7fc`](https://github.com/lblod/ember-rdfa-editor/commit/1b7a7fca607c41cbbfb9477cde19c30d710be2e1) Thanks [@dkozickis](https://github.com/dkozickis)! - GN-4820: Preprocess RDFa on paste
-
   - Preprocess RDFa on paste
   - Expose `recreateUuidsOnPaste` plugin `key` to allow overriding the plugin
 
@@ -690,7 +766,6 @@
 - [#1148](https://github.com/lblod/ember-rdfa-editor/pull/1148) [`d310875`](https://github.com/lblod/ember-rdfa-editor/commit/d310875b4d17465e3ff79acafc92420c4fb95c43) Thanks [@elpoelma](https://github.com/elpoelma)! - Add option to provide `EditorRange` object to `setHtmlContent` method in order to only replace a part of the document
 
 - [#1151](https://github.com/lblod/ember-rdfa-editor/pull/1151) [`4d0e54c`](https://github.com/lblod/ember-rdfa-editor/commit/4d0e54ca52268a6eac653f006ee9300ce97e2f56) Thanks [@elpoelma](https://github.com/elpoelma)! - Clean-up debug-tools component:
-
   - Usage of new `codemirror` modifier
   - Removal of `xml` support
 
@@ -704,7 +779,6 @@
 - [#1170](https://github.com/lblod/ember-rdfa-editor/pull/1170) [`16e083e`](https://github.com/lblod/ember-rdfa-editor/commit/16e083eb0a401c47d30581e5d701db3206b82902) Thanks [@dkozickis](https://github.com/dkozickis)! - GN-4707: Vertical alignment in table cells
 
 - [#1051](https://github.com/lblod/ember-rdfa-editor/pull/1051) [`0e15912`](https://github.com/lblod/ember-rdfa-editor/commit/0e15912922f5bdcd6f9a46a1fc396626d38f0402) Thanks [@elpoelma](https://github.com/elpoelma)! - Add option for the document (top) node to be parsed using `parseDOM` parse-rules. When setting the content of a document, using either `setHTMLContent` or `initialize`, three options are possible:
-
   - The `topNode` (often `doc`) has no parse-rules: a default node of type `topNode` is created (without any attributes). The html provided to `setHTMLContent` or `intialize` is parsed as its content.
   - The `topNode` has 1 or more parse-rules: the parser searches the provided html for a node that matches a parse-rule of the `topNode`.
     - If a node is found: the node is parsed as the `topNode` and its content is parsed as the `topNode` content
@@ -716,7 +790,6 @@
   For more information, check out [A new approach to handle RDFa in documents](https://github.com/lblod/ember-rdfa-editor/blob/9c32a9dea0da13df4092c39d9a092ba0803a3f42/README.md#experimental-a-new-approach-to-handle-rdfa-in-documents)
 
   #### Deprecations
-
   - Passing the `keyMapOptions` argument to the `RdfaEditor` is deprecated. The behaviour of `selectBlockRdfaNode` is included by default. This feature will be removed in the next major release.
   - The `rdfaAttrs` constant is deprecated, use the `rdfaAttrSpec` function instead
   - `inline_rdfa` is deprecated, use `inlineRdfaWithConfig` instead
@@ -777,7 +850,6 @@
 - [#1148](https://github.com/lblod/ember-rdfa-editor/pull/1148) [`d310875`](https://github.com/lblod/ember-rdfa-editor/commit/d310875b4d17465e3ff79acafc92420c4fb95c43) Thanks [@elpoelma](https://github.com/elpoelma)! - Add option to provide `EditorRange` object to `setHtmlContent` method in order to only replace a part of the document
 
 - [#1151](https://github.com/lblod/ember-rdfa-editor/pull/1151) [`4d0e54c`](https://github.com/lblod/ember-rdfa-editor/commit/4d0e54ca52268a6eac653f006ee9300ce97e2f56) Thanks [@elpoelma](https://github.com/elpoelma)! - Clean-up debug-tools component:
-
   - Usage of new `codemirror` modifier
   - Removal of `xml` support
 
@@ -791,7 +863,6 @@
 - [#1170](https://github.com/lblod/ember-rdfa-editor/pull/1170) [`16e083e`](https://github.com/lblod/ember-rdfa-editor/commit/16e083eb0a401c47d30581e5d701db3206b82902) Thanks [@dkozickis](https://github.com/dkozickis)! - GN-4707: Vertical alignment in table cells
 
 - [#1051](https://github.com/lblod/ember-rdfa-editor/pull/1051) [`0e15912`](https://github.com/lblod/ember-rdfa-editor/commit/0e15912922f5bdcd6f9a46a1fc396626d38f0402) Thanks [@elpoelma](https://github.com/elpoelma)! - Add option for the document (top) node to be parsed using `parseDOM` parse-rules. When setting the content of a document, using either `setHTMLContent` or `initialize`, three options are possible:
-
   - The `topNode` (often `doc`) has no parse-rules: a default node of type `topNode` is created (without any attributes). The html provided to `setHTMLContent` or `intialize` is parsed as its content.
   - The `topNode` has 1 or more parse-rules: the parser searches the provided html for a node that matches a parse-rule of the `topNode`.
     - If a node is found: the node is parsed as the `topNode` and its content is parsed as the `topNode` content
@@ -803,7 +874,6 @@
   For more information, check out [A new approach to handle RDFa in documents](https://github.com/lblod/ember-rdfa-editor/blob/9c32a9dea0da13df4092c39d9a092ba0803a3f42/README.md#experimental-a-new-approach-to-handle-rdfa-in-documents)
 
   #### Deprecations
-
   - Passing the `keyMapOptions` argument to the `RdfaEditor` is deprecated. The behaviour of `selectBlockRdfaNode` is included by default. This feature will be removed in the next major release.
   - The `rdfaAttrs` constant is deprecated, use the `rdfaAttrSpec` function instead
   - `inline_rdfa` is deprecated, use `inlineRdfaWithConfig` instead
@@ -918,7 +988,6 @@
 ### Major Changes
 
 - [#1051](https://github.com/lblod/ember-rdfa-editor/pull/1051) [`0e159129`](https://github.com/lblod/ember-rdfa-editor/commit/0e15912922f5bdcd6f9a46a1fc396626d38f0402) Thanks [@elpoelma](https://github.com/elpoelma)! - Add option for the document (top) node to be parsed using `parseDOM` parse-rules. When setting the content of a document, using either `setHTMLContent` or `initialize`, three options are possible:
-
   - The `topNode` (often `doc`) has no parse-rules: a default node of type `topNode` is created (without any attributes). The html provided to `setHTMLContent` or `intialize` is parsed as its content.
   - The `topNode` has 1 or more parse-rules: the parser searches the provided html for a node that matches a parse-rule of the `topNode`.
     - If a node is found: the node is parsed as the `topNode` and its content is parsed as the `topNode` content
@@ -1165,7 +1234,6 @@
   ```
 
 - [#1064](https://github.com/lblod/ember-rdfa-editor/pull/1064) [`35c4ca44`](https://github.com/lblod/ember-rdfa-editor/commit/35c4ca44d0d71a98c31ecebd9019174457970e87) Thanks [@elpoelma](https://github.com/elpoelma)! - Add shortcuts for setting the alignment of paragraphs/headers:
-
   - `Ctrl`+`Shift`+`L`: left align
   - `Ctrl`+`Shift`+`E`: center align
   - `Ctrl`+`Shift`+`R`: right align
@@ -1208,7 +1276,6 @@
 ### Patch Changes
 
 - [#1053](https://github.com/lblod/ember-rdfa-editor/pull/1053) [`c2e4e2a1`](https://github.com/lblod/ember-rdfa-editor/commit/c2e4e2a1e13c35ff99646f01eab0e236c68a4867) Thanks [@dkozickis](https://github.com/dkozickis)! - GN-4621: Add styling for selected table cell
-
   - `.selectedCell` to style the cell that is selected. `.selectedCell` comes from the `prosemirror-tables` plugin.
   - `::selection` to hide the selection on the text inside the cell.
 
@@ -1233,7 +1300,6 @@
 - [#1044](https://github.com/lblod/ember-rdfa-editor/pull/1044) [`0a2472c1`](https://github.com/lblod/ember-rdfa-editor/commit/0a2472c1e459091f10cb5cffdbe44e7b14bd2df3) Thanks [@elpoelma](https://github.com/elpoelma)! - Add shift+enter handler to `embeddedEditorBaseKeymap`
 
 - [#1027](https://github.com/lblod/ember-rdfa-editor/pull/1027) [`8d310cf1`](https://github.com/lblod/ember-rdfa-editor/commit/8d310cf1a3421935a858ad8a9483c4f83c4d7f66) Thanks [@elpoelma](https://github.com/elpoelma)! - Introduction of a custom gap-cursor plugin containing several fixes compared to the original version:
-
   - The click handler has been replaced by a mousedown handler in order to intercept a click event earlier
   - The types of the GapCursor class have been fixed
   - Addition of a fix when resolving the position returned by view.posAtCoords.
@@ -1301,7 +1367,6 @@
 ### Minor Changes
 
 - [#1027](https://github.com/lblod/ember-rdfa-editor/pull/1027) [`8d310cf1`](https://github.com/lblod/ember-rdfa-editor/commit/8d310cf1a3421935a858ad8a9483c4f83c4d7f66) Thanks [@elpoelma](https://github.com/elpoelma)! - Introduction of a custom gap-cursor plugin containing several fixes compared to the original version:
-
   - The click handler has been replaced by a mousedown handler in order to intercept a click event earlier
   - The types of the GapCursor class have been fixed
   - Addition of a fix when resolving the position returned by view.posAtCoords.
@@ -1413,7 +1478,6 @@
 - [#955](https://github.com/lblod/ember-rdfa-editor/pull/955) [`0f6da258`](https://github.com/lblod/ember-rdfa-editor/commit/0f6da25859f9d34be8ebce28b998f1bda192d9ad) Thanks [@elpoelma](https://github.com/elpoelma)! - GN-4479: allow for option to pass nodeviews object to embedded-editor component
 
 - [#962](https://github.com/lblod/ember-rdfa-editor/pull/962) [`49e52fb4`](https://github.com/lblod/ember-rdfa-editor/commit/49e52fb4c7cb8696898372e0d31df88036b56f2e) Thanks [@x-m-el](https://github.com/x-m-el)! - - Fix `leave-on-enter-key` modifier to always use the correct position to leave a node from.
-
   - Add `leave-with-arrow-keys` modifier, which can be used to leave an node's element with the left and right arrow key.
   - Add `select-node-on-click` modifier, which can be used to select a given node when clicking the bound element.